+++ front/php/action_product.php
@@ -24,6 +24,7 @@
         $stmt = mysqli_query($conn, $request);
             if($stmt){
                 move_uploaded_file($pro_img_tmp,$pro_image_folder);
+                header('location: accueil_admin.php');
             }else{
                 die("Échec de la connexion : " . $connexion->connect_error);
             }
@@ -82,7 +83,7 @@ if(isset($_POST['modifier_pro'])){
 
 <!-- ------------------------------Supprimer produit------------------------------------- -->
 <?php
-    session_start();
+if (isset($_GET['delete'])) {
     $id_pro = $_GET['delete'];
     $request = "DELETE FROM products WHERE reference = '$id_pro' ";
     $stmt = mysqli_query($conn,$request);  
@@ -91,4 +92,5 @@ if(isset($_POST['modifier_pro'])){
         } else {
         die("Échec : " . mysqli_error($conn));
         }
+}
   ?>
